ContentMachine/src/requirements.txt
numpy==1.24.2
openai==0.19.0
google-cloud-secret-manager==2.16.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
openai-whisper==20230314
dropbox==11.36.0
oauthlib==3.2.2 
requests-oauthlib==1.3.1 
tweepy==4.14.0
pyrebase4==4.6.0 
audioread==3.0.0
google-api-python-client==2.77.0
google-auth-oauthlib==1.0.0
praw==7.7.0
requests-toolbelt==0.10.1
Flask==2.3.2
Flask-Cors==3.0.10
isort==5.12.0
Flask[async]==2.3.2
google-cloud-storage==2.0.0
firebase-admin==6.1.0
firebase==4.0.1

